extract graph setting editors
