+++ src/plugins/history/Graph/types.ts
@@ -1,6 +1,4 @@
 import { GraphConfig } from '@/plugins/history/types';
 import { PersistentWidget } from '@/store/dashboards';
 
-export interface HistoryItem extends PersistentWidget {
-  config: GraphConfig;
-}
+export type HistoryItem = PersistentWidget<GraphConfig>;
